Report a file size of zero if the gridfs object does not exists.
This is needed on setups with multiple nodes. While one node is creating the
file the 'gridfs object' is not there yet. So the other node cannot get the file
size.

Yes I could have stored the filesize in the 'Entry' object, but having it at two
places doesn't feel 'right'.